Integrate Trimmed fields - Moved pieces in dedicated Django module (filters / forms) Added some Admin pages.
