+++ src/ensembl/production/masterdb/admin.py
@@ -10,25 +10,16 @@
 #   See the License for the specific language governing permissions and
 #   limitations under the License.
 # TODO add uncheck all is_current when checking is_current
-import collections
 
-from django import forms
 from django.contrib import admin
 from django.contrib import messages
-from django.contrib.admin import SimpleListFilter
 
 from ensembl.production.djcore.admin import ProductionUserAdminMixin
+from ensembl.production.djcore.utils import flatten
+
+from .filters import IsCurrentFilter
+from .forms import AnalysisDescriptionForm, WebDataForm
 from .models import *
-
-
-def flatten(iter_obj):
-    result = []
-    for el in iter_obj:
-        if isinstance(iter_obj, collections.Iterable) and not isinstance(el, str):
-            result.extend(flatten(el))
-        else:
-            result.append(el)
-    return result
 
 
 class ProductionModelAdmin(ProductionUserAdminMixin):
@@ -67,64 +58,6 @@ class ProductionModelAdmin(ProductionUserAdminMixin):
         if issubclass(self.model, BaseTimestampedModel):
             list_display = list_display + ('modified_at',)
         return list_display
-
-
-class IsCurrentFilter(SimpleListFilter):
-    title = 'Is Current'
-
-    parameter_name = 'is_current'
-
-    def lookups(self, request, model_admin):
-        return (
-            (None, 'Yes'),
-            ('no', 'No'),
-            ('all', 'All'),
-        )
-
-    def choices(self, cl):
-        for lookup, title in self.lookup_choices:
-            yield {
-                'selected': self.value() == lookup,
-                'query_string': cl.get_query_string({
-                    self.parameter_name: lookup,
-                }, []),
-                'display': title,
-            }
-
-    def queryset(self, request, queryset):
-        if self.value() == 'no':
-            return queryset.filter(is_current=False)
-        elif self.value() is None:
-            return queryset.filter(is_current=True)
-
-
-class IsDisplayableFilter(SimpleListFilter):
-    title = 'Is Displayed'
-
-    parameter_name = 'displayable'
-
-    def lookups(self, request, model_admin):
-        return (
-            (None, 'Yes'),
-            ('no', 'No'),
-            ('all', 'All'),
-        )
-
-    def choices(self, cl):
-        for lookup, title in self.lookup_choices:
-            yield {
-                'selected': self.value() == lookup,
-                'query_string': cl.get_query_string({
-                    self.parameter_name: lookup,
-                }, []),
-                'display': title,
-            }
-
-    def queryset(self, request, queryset):
-        if self.value() == 'no':
-            return queryset.filter(displayable=False)
-        elif self.value() is None:
-            return queryset.filter(displayable=True)
 
 
 class ProductionTabularInline(admin.TabularInline):
@@ -202,33 +135,15 @@ class BioTypeAdmin(HasCurrentAdmin):
               ('created_by', 'created_at'),
               ('modified_by', 'modified_at')
               )
-    list_display = ('name', 'object_type', 'db_type', 'biotype_group', 'attrib_type', 'description', 'is_current','so_acc', 'so_term')
-    search_fields = ('name', 'object_type', 'db_type', 'biotype_group', 'attrib_type__name', 'description','so_acc', 'so_term')
-
-
-class AnalysisDescriptionForm(forms.ModelForm):
-    class Meta:
-        model = AnalysisDescription
-        exclude = ('created_at', 'modified_at')
-
-    web_data_label = forms.CharField(required=False,
-                                     label="WebData content (ReadOnly)",
-                                     widget=forms.Textarea(attrs={'rows': 10, 'cols': 40, 'class': 'vLargeTextField',
-                                                                  'readonly': 'readonly'}))
-
-    def __init__(self, *args, **kwargs):
-        super(AnalysisDescriptionForm, self).__init__(*args, **kwargs)
-        current = kwargs.get('instance')
-        if current:
-            self.fields['web_data_label'].initial = current.web_data.label if current.web_data else ''
-        if not current or not current.web_data:
-            self.fields['web_data_label'].widget.attrs.update({'style': 'display:None'})
+    list_display = (
+        'name', 'object_type', 'db_type', 'biotype_group', 'attrib_type', 'description', 'is_current', 'so_acc',
+        'so_term')
+    search_fields = (
+        'name', 'object_type', 'db_type', 'biotype_group', 'attrib_type__name', 'description', 'so_acc', 'so_term')
 
 
 @admin.register(AnalysisDescription)
 class AnalysisDescriptionAdmin(HasCurrentAdmin):
-
-
     form = AnalysisDescriptionForm
     fields = ('logic_name', 'description', 'display_label', 'web_data',
               'web_data_label',
@@ -259,23 +174,6 @@ class MetakeyAdmin(HasCurrentAdmin):
         if obj is not None and 'name' not in read_only_fields:
             read_only_fields += ['name', ]
         return read_only_fields
-
-
-class WebDataForm(forms.ModelForm):
-    class Meta:
-        model = WebData
-        fields = ('data', 'comment')
-        widgets = {
-            'data': forms.Textarea(attrs={'rows': 20, 'class': 'vLargeTextField'}),
-            'comment': forms.Textarea(attrs={'rows': 7, 'class': 'vLargeTextField'}),
-        }
-
-    def get_initial_for_field(self, field, field_name):
-        if field_name == 'data':
-            return json.dumps(self.initial.get('data', field.initial), sort_keys=True, indent=4) if self.initial.get(
-                'data', field.initial) is not None else ""
-        else:
-            return super().get_initial_for_field(field, field_name)
 
 
 @admin.register(WebData)
@@ -316,3 +214,19 @@ class MasterExternalDbAdmin(HasCurrentAdmin):
     search_fields = (
         'db_name', 'db_release', 'status', 'db_display_name', 'priority', 'type', 'secondary_db_name',
         'secondary_db_table')
+
+
+@admin.register(MasterMiscSet)
+class MasterMiscSetAdmin(admin.ModelAdmin):
+    list_display = ('misc_set_id', 'code', 'name', 'short_description')
+    readonly_fields = ('misc_set_id',)
+    fields = ('misc_set_id', 'code', 'name',
+              'description', 'max_length', 'is_current',
+              ('created_by', 'created_at'),
+              ('modified_by', 'modified_at'))
+    search_fields = ('name', 'description', 'code')
+
+
+@admin.register(MasterUnmappedReason)
+class MasterUnmappedReasonAdmin(admin.ModelAdmin):
+    list_display = ('unmapped_reason_id', 'summary_description')
